+++ app/helpers/contacts_helper.rb
@@ -2,8 +2,4 @@ module ContactsHelper
   def has_contacts?(contacts)
     contacts.present? && contacts.any?
   end
-
-  def format_category_name(category)
-    category.tr("_", " ").capitalize
-  end
 end