+++ db/migrate/20230113150412_add_organisation_to_contact.rb
@@ -1,0 +1,5 @@
+class AddOrganisationToContact < ActiveRecord::Migration[7.0]
+  def change
+    add_column :contacts, :organisation_name, :string
+  end
+end