+++ spec/requests/contacts_controller_spec.rb
@@ -49,7 +49,7 @@ RSpec.describe ContactsController, type: :request do
   end
 
   describe "#create" do
-    let(:project) { create(:conversion_project) }
+    let(:project) { create(:voluntary_conversion_project) }
     let(:project_id) { project.id }
     let(:mock_contact) { build(:contact) }
     let(:new_contact_name) { "Josephine Bloggs" }
@@ -79,7 +79,7 @@ RSpec.describe ContactsController, type: :request do
 
     context "when the contact is valid" do
       it "saves the contact and redirects to the index view with a success message" do
-        expect(subject).to redirect_to(project_contacts_path(project.id))
+        expect(subject).to redirect_to(conversions_voluntary_project_contacts_path(project))
         expect(request.flash[:notice]).to eq(I18n.t("contact.create.success"))
 
         expect(Contact.count).to be 1